// src/types/tieumuc.ts
export type TieuMuc = {
  cac_tieu_muc_con: [
    {
      diem_toi_da: number | string;
      mo_ta_tieu_muc_con: string;
      muc: number | string;
      so_tieu_muc_con: string;
      ten_tieu_muc_con: string;
      hidden: number;
      id_tieumuccon: string;
    },
  ];
  diem_toi_da: number;
  mo_ta_tieu_muc: string;
  so_tieu_muc: string;
  ten_tieu_muc: string;
  hidden: number;
  id_tieumuc: string;
};